requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
cryptography==39.0.0
factory-boy==3.2.1
google-auth==2.15.0
gunicorn==20.1.0
requests-mock==1.10.0
sentry-sdk==1.12.1
sendgrid==6.9.7
django-storages==1.13.2
discord-webhook==1.0.0